update lam config

+++ gerrit/lam/install_lam.sh
@@ -6,14 +6,18 @@ VER=${DOCKER_LAM_VER:-6.6}
 LAM_PKG=ldap-account-manager-${VER}.tar.bz2
 LAM_URL=http://prdownloads.sourceforge.net/lam/${LAM_PKG}?download
 LAM_DIR=${DOCKER_LAM_DIR:-"/wwwroot/lam"}
+LAM_PKG_DIR="/wwwroot"
 
+PHP_INI_DIR=$(php --ini | grep "php.ini" | grep -oE "/.*")
 PHP_CONF_DIR=$(php --ini | grep additional | grep -o "/.*/etc/php.*")
 FPM_CONF_DIR="/usr/local/etc/php-fpm.d"
 
 config_php() {
+    cp ${PHP_INI_DIR}/php.ini-production ${PHP_INI_DIR}/php.ini
+
     #config php session
     local php_cfg=${PHP_CONF_DIR}/extra.ini
-    local PHP_SESSION_DIR=${WEB_ROOT}/zentaopms/tmp/session
+    local PHP_SESSION_DIR=${LAM_DIR}/tmp/session
 cat > ${php_cfg} << EOF
 session.save_handler = "files"
 session.save_path    = "${PHP_SESSION_DIR}"
@@ -42,10 +46,10 @@ init_deps() {
     #locale-gen zh_CN.UTF-8
     sed -i 's@dl-cdn.alpinelinux.org@mirrors.aliyun.com@g' /etc/apk/repositories
     apk add --no-cache --virtual .fetch-deps \
-            openldap-dev zlib-dev \
-            gettext-dev perl
+            openldap-dev zlib-dev libzip-dev \
+            gettext-dev perl libzip make
     apk add --no-cache --virtual .build-deps \
-            coreutils make
+            coreutils
     export PHP_EXT="gettext zip ldap"
 	docker-php-ext-install -j$(nproc) ${PHP_EXT}
 }
@@ -56,14 +60,18 @@ clean() {
 }
 
 download_extract_lam() {
-    curl -L ${LAM_URL} -o /tmp/${LAM_PKG}
-    cd /tmp
-    tar xf /tmp/${LAM_PKG}
+    local lam=/tmp/${LAM_PKG}
+    curl -L ${LAM_URL} -o $lam
+    tar xf $lam -C ${LAM_PKG_DIR}
 }
 
-compile_lam() {
+install_lam() {
+    if [ -f "${LAM_DIR}/VERSION" ];then
+        exit 0
+    fi
+
     [ -d ${LAM_DIR} ] || install -d ${LAM_DIR}
-    local LAM=/tmp/${LAM_PKG%.tar.bz2}
+    local LAM=${LAM_PKG_DIR}/${LAM_PKG%.tar.bz2}
     cd ${LAM}
     ./configure \
         --with-httpd-user=www-data \
@@ -73,14 +81,14 @@ compile_lam() {
         --sysconfdir=${LAM_DIR}/etc
     make install
     cd -
-    cd ${LAM_DIR}/config
-    cp config.cfg.sample config.cfg
-    cp unix.conf.sample lam.conf
-    cd -
+
+    # cd ${LAM_DIR}/config
+    # cp config.cfg.sample config.cfg
+    # cp unix.conf.sample lam.conf
+    # cd -
 }
 
 init_deps
 config_php
 download_extract_lam
-compile_lam
 clean
